Signal Scripts: Improved Exception Handling
Updated the codes to improve exception handling by checking for existing
views before creating the dynamic views.

+++ Signal_Scripts/SQL_Procedures/view_average_num_sig.sql
@@ -8,7 +8,11 @@ CREATE PROCEDURE view_average_num_sig (IN SIG_NAME VARCHAR(10))
 	LANGUAGE SQLSCRIPT AS
 	--DEFAULT SCHEMA "MIMIC2V26"
 BEGIN
---EXEC 'DROP VIEW "MIMIC2V26"."AVG_NUM_SIG_' || :SIG_NAME || '"';
+DECLARE view_exists INT := 0;
+SELECT COUNT(*) INTO view_exists FROM "VIEWS" WHERE "SCHEMA_NAME" = 'MIMIC2V26' AND "VIEW_NAME" = 'AVG_NUM_SIG_'  || :SIG_NAME;
+IF :view_exists = 1 THEN
+	EXEC 'DROP VIEW "MIMIC2V26"."AVG_NUM_SIG_' || :SIG_NAME || '"';
+END IF;
 EXEC 'CREATE VIEW "MIMIC2V26"."AVG_NUM_SIG_' || :SIG_NAME || '" AS 
 ( SELECT
 	"t"."SUBJECT_ID",
@@ -19,8 +23,8 @@ EXEC 'CREATE VIEW "MIMIC2V26"."AVG_NUM_SIG_' || :SIG_NAME || '" AS
 			"MIMIC2V26"."wav_num_records"."SUBJECT_ID",
 			"MIMIC2V26"."wav_num_records"."RECORD_ID",
 			AVG("MIMIC2V26"."wav_num_sig_' || :SIG_NAME || '"."AMPLITUDE") AS "AVG_AMP"
- 			FROM "MIMIC2V26"."wav_num_sig_' || :SIG_NAME || '", "MIMIC2V26"."wav_num_records"
- 			WHERE "MIMIC2V26"."wav_num_records"."RECORD_ID" = "MIMIC2V26"."wav_num_sig_' || :SIG_NAME || '"."RECORD_ID"
+			FROM "MIMIC2V26"."wav_num_sig_' || :SIG_NAME || '", "MIMIC2V26"."wav_num_records"
+			WHERE "MIMIC2V26"."wav_num_records"."RECORD_ID" = "MIMIC2V26"."wav_num_sig_' || :SIG_NAME || '"."RECORD_ID"
 				AND "MIMIC2V26"."wav_num_sig_' || :SIG_NAME || '"."AMPLITUDE" > 0
 			GROUP BY "MIMIC2V26"."wav_num_records"."SUBJECT_ID", "MIMIC2V26"."wav_num_records"."RECORD_ID"
 		) AS "t",
